kernel: Only download keyring if missing

+++ mkosi.conf.d/01-kernel.sh
@@ -2,8 +2,10 @@
 [ "$1" = "final" ] || exit 0
 
 # Get the repository keyring key.
-mkdir -p /etc/apt/keyrings/
-curl -fsSL https://pkgs.zabbly.com/key.asc -o /etc/apt/keyrings/zabbly.asc
+if [ ! -e /etc/apt/keyrings/zabbly.asc ]; then
+    mkdir -p /etc/apt/keyrings/
+    curl -fsSL https://pkgs.zabbly.com/key.asc -o /etc/apt/keyrings/zabbly.asc
+fi
 
 # Add the repository.
 cat <<EOF > /etc/apt/sources.list.d/zabbly-kernel-stable.sources
